book_project_V2a/requirements.txt
python-dotenv
python-multipart
pymongo[srv]
fastapi
uvicorn
pytest
httpx
requests